+++ src/wurf/check_lock_cache_resolver.py
@@ -7,7 +7,7 @@ from .error import DependencyError
 class CheckLockCacheResolver(object):
     """Iterates through a list of resolvers until a path is resolved."""
 
-    def __init__(self, resolver, lock_cache, dependency):
+    def __init__(self, resolver, lock_cache_from, dependency):
         """Construct an instance.
 
         :param resolvers: A list of resolvers object for the available
@@ -16,7 +16,7 @@ class CheckLockCacheResolver(object):
         :param dependency: A Dependency instance.
         """
         self.resolver = resolver
-        self.lock_cache = lock_cache
+        self.lock_cache = lock_cache_from
         self.dependency = dependency
 
     def resolve(self):
@@ -31,17 +31,23 @@ class CheckLockCacheResolver(object):
                 dependency=self.dependency,
             )
 
-        if self.dependency.resolver == "git":
-            if self.lock_cache.check_sha1(dependency=self.dependency):
-                raise DependencyError(
-                    msg="Locked SHA1 mismatch.",
-                    dependency=self.dependency,
-                )
+        if self.lock_cache.check_sha1(dependency=self.dependency):
+            raise DependencyError(
+                msg=(
+                    "Locked dependency inconsistent with the "
+                    "dependency specified in the project."
+                ),
+                dependency=self.dependency,
+            )
+        # We need to call resolve before checking the content as we need to know
+        # where the content is located.
         path = self.resolver.resolve()
-        if self.dependency.resolver == "http":
-            if self.lock_cache.check_file_hash(dependency=self.dependency, path=path):
-                raise DependencyError(
-                    msg="Locked file hash mismatch.",
-                    dependency=self.dependency,
-                )
+        if self.lock_cache.check_content(dependency=self.dependency):
+            raise DependencyError(
+                msg=(
+                    "The content does not match the data "
+                    "when the dependency was locked."
+                ),
+                dependency=self.dependency,
+            )
         return path