+++ src/wurf/store_lock_path_resolver.py
@@ -4,20 +4,23 @@
 import os
 
 from .error import WurfError
+from .lock_path_cache import LockPathCache
 
 
 class StoreLockPathResolver(object):
-    def __init__(self, resolver, lock_cache, project_path, dependency):
+    def __init__(
+        self, resolver, lock_cache_to: LockPathCache, project_path, dependency
+    ):
         """Construct an instance.
 
         :param resolver: A resolver which will do the actual job
-        :param lock_cache: The lock cache to store the version information in.
+        :param lock_cache_to: The lock cache to store the version information in.
         :param project_path: The path to the project who's dependencies we are
            resolving as a string.
         :param dependency: A Dependency instance.
         """
         self.resolver = resolver
-        self.lock_cache = lock_cache
+        self.lock_cache_to: LockPathCache = lock_cache_to
         self.project_path = project_path
         self.dependency = dependency
 
@@ -41,14 +44,15 @@ class StoreLockPathResolver(object):
 
         lock_path = os.path.relpath(path=lock_path, start=self.project_path)
 
-        self.lock_cache.add_path(dependency=self.dependency, path=lock_path)
+        self.lock_cache_to.add_path(dependency=self.dependency, path=lock_path)
 
         return path
 
     def __check_path(self, path):
-
         child = os.path.realpath(path)
         parent = os.path.realpath(self.project_path)
 
         if not child.startswith(parent):
-            raise WurfError("Locked paths must be a within parent project")
+            raise WurfError(
+                f"Locked paths must be a within parent project {parent} {child}"
+            )